+++ pkg/kube-metrics/store.go
@@ -27,10 +27,10 @@ import (
 	metricsstore "k8s.io/kube-state-metrics/pkg/metrics_store"
 )
 
-// NewMetricsStores returns collections of metrics in the namespaces provided, per the api/kind resource.
+// NewNamespacedMetricsStores returns collections of metrics in the namespaces provided, per the api/kind resource.
 // The metrics are registered in the custom metrics.FamilyGenerator that needs to be defined.
-func NewMetricsStores(dclient dynamic.NamespaceableResourceInterface, namespaces []string, api string, kind string,
-	metricFamily []metric.FamilyGenerator) []*metricsstore.MetricsStore {
+func NewNamespacedMetricsStores(dclient dynamic.NamespaceableResourceInterface, namespaces []string,
+	api string, kind string, metricFamily []metric.FamilyGenerator) []*metricsstore.MetricsStore {
 	namespaces = deduplicateNamespaces(namespaces)
 	var stores []*metricsstore.MetricsStore
 	// Generate collector per namespace.
@@ -44,6 +44,30 @@ func NewMetricsStores(dclient dynamic.NamespaceableResourceInterface, namespaces
 	return stores
 }
 
+// NewClusterScopedMetricsStores returns collections of metrics per the api/kind resource.
+// The metrics are registered in the custom metrics.FamilyGenerator that needs to be defined.
+func NewClusterScopedMetricsStores(dclient dynamic.NamespaceableResourceInterface, api string, kind string,
+	metricFamily []metric.FamilyGenerator) []*metricsstore.MetricsStore {
+	var stores []*metricsstore.MetricsStore
+	// Generate collector per cluster scoped resources.
+	composedMetricGenFuncs := metric.ComposeMetricGenFuncs(metricFamily)
+	headers := metric.ExtractMetricFamilyHeaders(metricFamily)
+	store := metricsstore.NewMetricsStore(headers, composedMetricGenFuncs)
+	reflectorPerClusterScoped(context.TODO(), dclient, &unstructured.Unstructured{}, store)
+	stores = append(stores, store)
+
+	return stores
+}
+
+// NewMetricsStores returns collections of metrics in the namespaces provided, per the api/kind resource.
+// The metrics are registered in the custom metrics.FamilyGenerator that needs to be defined.
+//
+// Deprecated: Use NewNamespacedMetricsStores instead.
+func NewMetricsStores(dclient dynamic.NamespaceableResourceInterface, namespaces []string,
+	api string, kind string, metricFamily []metric.FamilyGenerator) []*metricsstore.MetricsStore {
+	return NewNamespacedMetricsStores(dclient, namespaces, api, kind, metricFamily)
+}
+
 func deduplicateNamespaces(ns []string) (list []string) {
 	keys := make(map[string]struct{})
 	for _, entry := range ns {
@@ -55,6 +79,17 @@ func deduplicateNamespaces(ns []string) (list []string) {
 	return list
 }
 
+func reflectorPerClusterScoped(
+	ctx context.Context,
+	dynamicInterface dynamic.NamespaceableResourceInterface,
+	expectedType interface{},
+	store cache.Store,
+) {
+	lw := clusterScopeListWatchFunc(dynamicInterface)
+	reflector := cache.NewReflector(&lw, expectedType, store, 0)
+	go reflector.Run(ctx.Done())
+}
+
 func reflectorPerNamespace(
 	ctx context.Context,
 	dynamicInterface dynamic.NamespaceableResourceInterface,
@@ -62,18 +97,30 @@ func reflectorPerNamespace(
 	store cache.Store,
 	ns string,
 ) {
-	lw := listWatchFunc(dynamicInterface, ns)
+	lw := namespacedListWatchFunc(dynamicInterface, ns)
 	reflector := cache.NewReflector(&lw, expectedType, store, 0)
 	go reflector.Run(ctx.Done())
 }
 
-func listWatchFunc(dynamicInterface dynamic.NamespaceableResourceInterface, namespace string) cache.ListWatch {
+func namespacedListWatchFunc(dynamicInterface dynamic.NamespaceableResourceInterface,
+	namespace string) cache.ListWatch {
 	return cache.ListWatch{
 		ListFunc: func(opts metav1.ListOptions) (runtime.Object, error) {
 			return dynamicInterface.Namespace(namespace).List(opts)
 		},
 		WatchFunc: func(opts metav1.ListOptions) (watch.Interface, error) {
 			return dynamicInterface.Namespace(namespace).Watch(opts)
+		},
+	}
+}
+
+func clusterScopeListWatchFunc(dynamicInterface dynamic.NamespaceableResourceInterface) cache.ListWatch {
+	return cache.ListWatch{
+		ListFunc: func(opts metav1.ListOptions) (runtime.Object, error) {
+			return dynamicInterface.List(opts)
+		},
+		WatchFunc: func(opts metav1.ListOptions) (watch.Interface, error) {
+			return dynamicInterface.Watch(opts)
 		},
 	}
 }